+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/layout/ListDividerItemDecoration.kt
@@ -4,8 +4,7 @@ import android.content.Context
 import androidx.recyclerview.widget.DividerItemDecoration
 import com.agileburo.anytype.core_ui.R
 
-class ListDividerItemDecoration(context: Context) :
-    DividerItemDecoration(context, VERTICAL) {
+class ListDividerItemDecoration(context: Context) : DividerItemDecoration(context, VERTICAL) {
 
     init {
         context.getDrawable(R.drawable.divider)?.let {