+++ library-page-icon-picker-widget/src/main/java/com/agileburo/anytype/library_page_icon_picker_widget/ui/ActionMenuDivider.kt
@@ -1,0 +1,21 @@
+package com.agileburo.anytype.library_page_icon_picker_widget.ui
+
+import android.graphics.Canvas
+import android.graphics.drawable.Drawable
+import androidx.recyclerview.widget.RecyclerView
+
+class ActionMenuDivider(private val divider: Drawable) : RecyclerView.ItemDecoration() {
+    override fun onDraw(c: Canvas, parent: RecyclerView, state: RecyclerView.State) {
+        val dividerLeft = parent.paddingLeft
+        val dividerRight = parent.width - parent.paddingRight
+        val count = parent.childCount
+        for (i in 0..count - 2) {
+            val child = parent.getChildAt(i)
+            val params = child.layoutParams as RecyclerView.LayoutParams
+            val dividerTop = child.bottom + params.bottomMargin
+            val dividerBottom = dividerTop + divider.intrinsicHeight
+            divider.setBounds(dividerLeft, dividerTop, dividerRight, dividerBottom)
+            divider.draw(c)
+        }
+    }
+}