+++ presentation/src/main/java/com/agileburo/anytype/presentation/page/PageViewModel.kt
@@ -1627,7 +1627,7 @@ class PageViewModel(
     }
 
     fun onPageIconClicked() {
-        dispatch(Command.OpenPagePicker(context))
+        dispatch(Command.OpenDocumentIconActionMenu(context))
     }
 
     private fun onFileClicked(id: String) {